backend/requirements.txt
# FixBot Backend Requirements
# Python 3.11+

# ============================================
# Web Framework
# ============================================
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6
websockets>=12.0

# ============================================
# Environment & Configuration
# ============================================
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# ============================================
# HTTP Clients
# ============================================
httpx>=0.28.0
aiohttp>=3.11.0
requests>=2.32.0

# ============================================
# DigitalOcean
# ============================================
python-digitalocean==1.17.0

# ============================================
# Infrastructure as Code
# ============================================
python-terraform==0.10.1

# ============================================
# Monitoring & Metrics
# ============================================
prometheus-client==0.19.0
prometheus-api-client==0.5.3

# ============================================
# Data Processing
# ============================================
numpy>=1.26.0
pandas>=2.2.0

# ============================================
# Async & Task Queue
# ============================================
# asyncio is built into Python 3.x, no need to install separately
aiofiles==23.2.1

# ============================================
# Logging & Observability
# ============================================
structlog==23.2.0
python-json-logger==2.0.7

# ============================================
# Testing
# ============================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-httpx>=0.30.0

# ============================================
# Code Quality
# ============================================
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# ============================================
# SSH & Remote Execution
# ============================================
paramiko==3.4.0
fabric==3.2.2

# ============================================
# Utilities
# ============================================
python-dateutil==2.8.2
pytz==2023.3
# uuid is built into Python, no need to install separately
tenacity==8.2.3

# ============================================
# CORS
# ============================================
# python-cors is not a valid package, CORS is handled by FastAPI middleware

# ============================================
# Optional: Dashboard (Streamlit)
# ============================================
streamlit>=1.40.0
plotly>=5.18.0
